requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
discord.py==2.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
multidict==6.1.0
outcome==1.3.0.post0
propcache==0.2.1
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
PySocks==1.7.1
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink==7.1.1
trio==0.28.0
trio-websocket==0.11.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
